Mobile planner: start out with some QML
This creates the UI part of a mobile planner, with:
-Dive planner page with listviews for cylinders and dive planner points
-Overlay to add/edit cylinder details
-Overlay to add/edit planner points
-Overlay to edit dive planner preferences

This doesn't work at all to plan dives. It needs to be connected to the
c/c++/Qt code that calculates the plan, sets preferences, etc. There are dummy
models for cylinders and planner points so that something is shown. These need
to be replaced by the actual Qt models.

Signed-off-by: Rick Walsh <rickmwalsh@gmail.com>

on '1b8b1120'.